% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_enum_method.R
\name{apply_enum_method}
\alias{apply_enum_method}
\title{@title TEMPORARRY DOCUMENTATION}
\usage{
apply_enum_method(enum_value, method, ...)
}
\description{
@description TEMPORARY DOCUMENTATION
}
\details{
@export
}
