% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessing_enum_values.R
\name{$<-.enum_type}
\alias{$<-.enum_type}
\title{Get subset of values of enum_type}
\usage{
\method{$}{enum_type}(enum_type, index) <- value
}
\arguments{
\item{enum_type}{enum_type}

\item{index}{index}
}
\description{
placeholder
}
