% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum_index.R
\name{get_enum_index}
\alias{get_enum_index}
\title{@title Get Enum Index}
\usage{
get_enum_index(enum_value)
}
\description{
@description Extracts `index` attribute value of passed `enum_value`.
}
\details{
@param enum_value object of class `enum_value`

@return integer index of

@export
}
