% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_enum_method.R
\name{get_enum_method}
\alias{get_enum_method}
\title{@title TEMPORARRY DOCUMENTATION}
\usage{
get_enum_method(enum_val, method)
}
\description{
@description TEMPORARY DOCUMENTATION
}
\details{
@export
}
