% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_enum_value.R
\name{match_enum_value}
\alias{match_enum_value}
\title{TEMPORARRY DOCUMENTATION}
\usage{
match_enum_value(x, table, nomatch = NA_integer_, incomparables = NULL)
}
\description{
TEMPORARY DOCUMENTATION
}
