% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.enum_type.R
\name{print.enum_type}
\alias{print.enum_type}
\title{Print enum_type}
\usage{
\method{print}{enum_type}(enum_type)
}
\arguments{
\item{enum_type}{Object of class enum_type}
}
\description{
Pretty prints enum_type object. Overloads default print.
}
