% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.enum_value.R
\name{print.enum_value}
\alias{print.enum_value}
\title{@title TEMPORARRY DOCUMENTATION}
\usage{
\method{print}{enum_value}(enum_value)
}
\description{
@description TEMPORARY DOCUMENTATION
}
\details{
@export
}
