% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.enum_type.R
\name{as.enum_type}
\alias{as.enum_type}
\title{Coerce to an enum type}
\usage{
as.enum_type(x)
}
\arguments{
\item{x}{any R object}
}
\value{
Object of class \code{\link{enum_type}}
}
\description{
Functions to check if an object is an enum type, or coerce it if possible.
}
