% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.enum_type.R
\name{as.enum_type}
\alias{as.enum_type}
\title{@title Coerce to an enum type}
\usage{
as.enum_type(x)
}
\description{
@description Functions to check if an object is an enum type, or coerce it if possible.
}
\details{
@param x any R object

@return Object of class \code{\link{enum_type}}

@export
}
