% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum_type.R
\name{enum_type}
\alias{enum_type}
\title{@title Enum type}
\usage{
enum_type(type_name, values)
}
\description{
@description Creates new enum type.
}
\details{
@param type_name Name of the type. Type will be referenced by this value
 @param values Possible values of print

@export
}
