% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_enum.R
\name{sample_enum}
\alias{sample_enum}
\title{@title Sample n values from an enum type}
\usage{
sample_enum(enum_type, size, replace = FALSE, weight = NULL)
}
\description{
@description This is a wrapper around [sample()] to make selecting random enum values easier.
}
\details{
@param enum_type Object that passes [is.enum_type()] check.
 @param size A non-negative integer giving number of items to choose.
 @param replace Should sampling be with replacement?
 @param weight Sampling weights. This must evaluate to a vector of non-negative numbers the same
 length as the input. Weights are automatically standardised to sum to 1.

@return A list of enum values of length \code{size} drawn from \code{enum_type}'s values.

@export
}
